refactor: split use-supplier-profile into focused single-responsibility modules (#61)
- Extract company CRUD into hooks/use-supplier-companies.ts
- Extract product CRUD + dialog state into hooks/use-supplier-products.ts
- Extract derived inventory metrics into hooks/use-inventory-stats.ts
- Extract pure storage helpers into lib/supplier-profile/storage.ts
- Compose all sub-hooks in use-supplier-profile.ts (612 -> 132 lines)
- Zero API surface changes to the consumer
